etl/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ohio-fit-etl"
version = "0.1.0"
description = "Ohio FIT ETL to fetch, normalize, and aggregate summarized financials"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "pyarrow>=16.0",
  "openpyxl>=3.1",
  "requests>=2.32",
  "pydantic>=2.7",
  "typer>=0.12",
  "ruamel.yaml>=0.18",
]

[project.scripts]
ohio-fit-etl = "ohio_fit_etl.cli:app"
